# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: API for Clients
  description: API for make CRUD operations of Client & Person.
  version: 1.0.0
  contact:
    name: José Obando
    email: jose.obando_0001@hotmail.com
servers:
  - url: http://localhost:8080/api/v1/
    description: Development
tags:
  - name: Customer
    description: Customer Controller
  - name: Person
    description: Person Controller
paths:
  /customers:
    get:
      tags:
        - Customer
      operationId: GetCustomerByFilter
      summary: Get list of clients
      description: List of clients.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientResponse'
    post:
      tags:
        - Customer
      summary: Create new client
      operationId: PostCustomer
      description: Create new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientRequest'
      responses:
        '201':
          description: Client created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'

  /customers/{userId}:
    get:
      tags:
        - Customer
      summary: Get a customer by ID
      operationId: GetCustomerById
      description: Get a customer by ID.
      parameters:
        - name: userId
          in: path
          required: true
          description: User Id
          schema:
            type: string
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '404':
          description: User not found

    put:
      tags:
        - Customer
      summary: Update a customer
      operationId: PutCustomer
      description: Update a customer.
      parameters:
        - name: userId
          in: path
          required: true
          description: User Id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientRequest'
      responses:
        '200':
          description: Client updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '404':
          description: Client not found

    delete:
      tags:
        - Customer
      summary: Delete customer
      operationId: DeleteCustomer
      description: Delete customer
      parameters:
        - name: userId
          in: path
          required: true
          description: UserId
          schema:
            type: string
      responses:
        '204':
          description: Client deleted
        '404':
          description: Client not found

  /person:
    get:
      tags:
        - Person
      operationId: GetPersonByFilter
      summary: Get list of people
      description: List of people.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PersonResponse'
    post:
      tags:
        - Person
      summary: Create new person
      operationId: PostPerson
      description: Create new person
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonRequest'
      responses:
        '201':
          description: Person created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponse'

components:
  schemas:
    ClientRequest:
      type: object
      required:
        - names
        - address
      properties:
        names:
          type: string
          minLength: 1
          maxLength: 100
          description: Names for client
          example: Jose Lema
        address:
          type: string
          minLength: 1
          maxLength: 200
          description: Address for client
          example: Otavalo sn y principal
        phone:
          type: string
          pattern: '^\d{10}$'
          description: Phone number for client
          example: 0986492314
        password:
          type: string
          minLength: 6
          maxLength: 20
          description: Password for client
          example: "*******"
        status:
          type: boolean
          description: Status for client
          example: true

    ClientResponse:
      type: object
      properties:
        userId:
          type: string
          description: Unique identifier for the client
        names:
          type: string
          description: Names of the client
        address:
          type: string
          description: Address of the client
        phone:
          type: string
          description: Phone number of the client
        password:
          type: string
          description: Password for client (might be hidden in responses)
        status:
          type: boolean
          description: Status for client

    PersonRequest:
      type: object
      required:
        - name
        - gender
        - identification
        - address
        - phone
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Name of the person
        gender:
          type: string
          enum:
            - MALE
            - FEMALE
            - OTHER
          description: Gender of the person
        identification:
          type: string
          minLength: 1
          maxLength: 20
          description: Identification number or document of the person
        address:
          type: string
          minLength: 1
          maxLength: 200
          description: Address of the person
        phone:
          type: string
          pattern: '^\d{10}$'
          description: Phone number of the person

    PersonResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the person
        name:
          type: string
          description: Name of the person
        gender:
          type: string
          enum:
            - MALE
            - FEMALE
            - OTHER
          description: Gender of the person
        identification:
          type: string
          description: Identification number or document of the person
        address:
          type: string
          description: Address of the person
        phone:
          type: string
          description: Phone number of the person